move flag parsing to pkg
